requirements.txt
Telethon==1.34.0
APScheduler==3.10.4
python-dotenv==1.0.0
pytz==2024.1
